+++ server/router/routes/users.js
@@ -1,4 +1,4 @@
-// This handles retrieving of users
+// This handles retrieving of Users
 
 // Includes Express
 var express = require('express');
@@ -9,6 +9,7 @@ var moment = require('moment');
 var _ = require('underscore');
 var color = require('cli-color');
 
+// Access to Database
 var db = require('../../database');
 var Users = db.users;
 
@@ -34,10 +35,6 @@ router.post('/', function (req,res) {
 
 		// If there's an error, log it and return to user
 		if (err) {
-			// Nice log message on your end, so that you can see what happened
-			console.log('Couldn\'t create new user at ' + color.red(time) + ' by ' + color.red(body.email) + ' because of: ' + err);
-
-			// send the error
 			res.status(500).json({
 				'message': 'Internal server error from signing up new user. Please contact support@yourproject.com.'
 			});
@@ -45,9 +42,6 @@ router.post('/', function (req,res) {
 
 		// If the user doesn't exist, create one
 		if (!user) {
-			console.log('Creating a new user at ' + color.green(time) + ' with the email: ' + color.green(body.email));
-
-			// setup the new user
 			var newUser = new Users({
 				firstname: body.firstName,
 				lastname: body.lastName,
@@ -55,23 +49,12 @@ router.post('/', function (req,res) {
 				password: body.confirmPwd
 			});
 
-			console.log("firstname: " + body.firstName);
-			console.log("lastname: " + body.lastName);
-			console.log("email: " + body.email);
-			console.log("password: " + body.confirmPwd);
-
-			// save the user to the database
 			newUser.save(function (err, savedUser, numberAffected) {
-
 				if (err) {
-					console.log('Problem saving the user ' + color.yellow(body.email) + ' due to ' + err);
 					res.status(500).json({
 						'message': 'Database error trying to sign up.  Please contact support@yourproject.com.'
 					});
 				}
-
-				// Log success and send the filtered user back
-				console.log('Successfully created new user: ' + color.green(body.email));
 
 				res.status(201).json({
 					'message': 'Successfully created new user',
@@ -83,10 +66,6 @@ router.post('/', function (req,res) {
 
 		// If the user already exists...
 		if (user) {
-
-			console.log('User ' + color.red(body.email) + ' already exists.');
-
-
 			res.status(409).json({
 				'message': body.email + ' already exists!'
 			});